Replace original EntityTrackingFlow with simplified version
This commit:
- Replaces the original entity_tracking_flow.py with the simplified version
- Renames the test files to match the project's conventions
- Updates import references in test files
- Verifies all tests pass with the simplified implementation
