+++ tests/flows/entity_tracking_flow_test.py
@@ -1,218 +1,211 @@
-"""Tests for the Entity Tracking flow."""
+"""Tests for the simplified EntityTrackingFlow implementation."""
 
-from datetime import datetime, timezone
-from unittest.mock import Mock, patch, MagicMock
+from datetime import datetime
+from unittest.mock import Mock, MagicMock
 
 import pytest
+from sqlmodel import Session
 
 from local_newsifier.flows.entity_tracking_flow import EntityTrackingFlow
-from local_newsifier.models.state import EntityTrackingState, EntityBatchTrackingState, EntityDashboardState, EntityRelationshipState, TrackingStatus
+from local_newsifier.models.state import EntityTrackingState, TrackingStatus
 from local_newsifier.services.entity_service import EntityService
+from local_newsifier.tools.entity_tracker_service import EntityTracker
+from local_newsifier.tools.extraction.entity_extractor import EntityExtractor
+from local_newsifier.tools.analysis.context_analyzer import ContextAnalyzer
+from local_newsifier.tools.resolution.entity_resolver import EntityResolver
 
 
-@patch("local_newsifier.flows.entity_tracking_flow.EntityExtractor")
-@patch("local_newsifier.flows.entity_tracking_flow.ContextAnalyzer")
-@patch("local_newsifier.flows.entity_tracking_flow.EntityResolver")
-@patch("local_newsifier.flows.entity_tracking_flow.EntityTracker")
-@patch("local_newsifier.flows.entity_tracking_flow.EntityService")
-def test_entity_tracking_flow_init(
-    mock_entity_service_class, 
-    mock_tracker_class,
-    mock_resolver_class, 
-    mock_context_analyzer_class, 
-    mock_extractor_class
-):
-    """Test initializing the entity tracking flow with defaults."""
-    # Setup mocks
+def test_entity_tracking_flow_with_explicit_dependencies():
+    """Test with explicitly provided dependencies."""
+    # Arrange - Create mocks
     mock_entity_service = Mock(spec=EntityService)
-    mock_entity_service_class.return_value = mock_entity_service
+    mock_entity_tracker = Mock(spec=EntityTracker)
+    mock_entity_extractor = Mock(spec=EntityExtractor)
+    mock_context_analyzer = Mock(spec=ContextAnalyzer)
+    mock_entity_resolver = Mock(spec=EntityResolver)
     
-    mock_tracker = Mock()
-    mock_tracker_class.return_value = mock_tracker
+    # Setup mock responses
+    mock_result_state = Mock(spec=EntityTrackingState)
+    mock_result_state.status = TrackingStatus.SUCCESS
+    mock_result_state.entities = [
+        {
+            "original_text": "John Doe",
+            "canonical_name": "John Doe",
+            "canonical_id": 1,
+            "context": "John Doe visited the city.",
+            "sentiment_score": 0.5,
+            "framing_category": "neutral"
+        }
+    ]
+    mock_entity_service.process_article_with_state.return_value = mock_result_state
     
-    mock_extractor = Mock()
-    mock_extractor_class.return_value = mock_extractor
-    
-    mock_analyzer = Mock()
-    mock_context_analyzer_class.return_value = mock_analyzer
-    
-    mock_resolver = Mock()
-    mock_resolver_class.return_value = mock_resolver
-    
-    # Initialize flow
-    flow = EntityTrackingFlow()
-    
-    # Verify service and tools were created
-    assert flow.entity_service is not None
-    assert flow.session is None
-    assert flow._entity_tracker is not None
-    assert flow._entity_extractor is not None
-    assert flow._context_analyzer is not None
-    assert flow._entity_resolver is not None
+    # Create test state
+    state = EntityTrackingState(
+        article_id=1,
+        content="John Doe visited the city.",
+        title="Test Article",
+        published_at=datetime(2025, 1, 1)
+    )
 
-
-def test_entity_tracking_flow_init_with_dependencies():
-    """Test initializing the entity tracking flow with provided dependencies."""
-    # Setup mocks
-    mock_entity_service = Mock(spec=EntityService)
-    mock_entity_tracker = Mock()
-    mock_entity_extractor = Mock()
-    mock_context_analyzer = Mock()
-    mock_entity_resolver = Mock()
-    mock_session_factory = Mock()
-    mock_session = Mock()
-    
-    # Initialize flow with mock dependencies
+    # Act - Initialize flow with explicit dependencies
     flow = EntityTrackingFlow(
         entity_service=mock_entity_service,
         entity_tracker=mock_entity_tracker,
         entity_extractor=mock_entity_extractor,
         context_analyzer=mock_context_analyzer,
-        entity_resolver=mock_entity_resolver,
-        session_factory=mock_session_factory,
-        session=mock_session
+        entity_resolver=mock_entity_resolver
     )
     
-    # Verify dependencies were used
-    assert flow.entity_service is mock_entity_service
-    assert flow._entity_tracker is mock_entity_tracker
-    assert flow._entity_extractor is mock_entity_extractor
-    assert flow._context_analyzer is mock_context_analyzer
-    assert flow._entity_resolver is mock_entity_resolver
-    assert flow.session is mock_session
+    # Call process method
+    result_state = flow.process(state)
+
+    # Assert - Verify correct dependencies were used
+    mock_entity_service.process_article_with_state.assert_called_once_with(state)
+    assert result_state is mock_result_state
+    assert result_state.status == TrackingStatus.SUCCESS
+    assert len(result_state.entities) == 1
+    assert result_state.entities[0]["original_text"] == "John Doe"
 
 
-def test_process_method():
-    """Test the process method."""
-    # Setup mocks
+def test_entity_tracking_flow_with_container():
+    """Test flow using container-based dependency resolution."""
+    # Arrange - Create mock container
+    mock_container = Mock()
     mock_entity_service = Mock(spec=EntityService)
-    mock_state = Mock(spec=EntityTrackingState)
+    
+    # Configure container to return mock service
+    mock_container.get.side_effect = lambda name: {
+        "entity_service": mock_entity_service,
+        "entity_tracker_tool": Mock(spec=EntityTracker),
+        "entity_extractor_tool": Mock(spec=EntityExtractor),
+        "context_analyzer_tool": Mock(spec=ContextAnalyzer),
+        "entity_resolver_tool": Mock(spec=EntityResolver),
+    }.get(name)
+    
+    # Setup mock response
     mock_result_state = Mock(spec=EntityTrackingState)
     mock_entity_service.process_article_with_state.return_value = mock_result_state
     
-    # Initialize flow
-    flow = EntityTrackingFlow(entity_service=mock_entity_service)
+    # Create test state
+    state = EntityTrackingState(
+        article_id=1,
+        content="Test content",
+        title="Test Article",
+        published_at=datetime(2025, 1, 1)
+    )
+    
+    # Act - Initialize flow with container
+    flow = EntityTrackingFlow(container=mock_container)
+    
+    # Call process method - should use container to get dependencies
+    result = flow.process(state)
+    
+    # Assert - Verify container was used to get dependencies
+    mock_container.get.assert_any_call("entity_service")
+    mock_entity_service.process_article_with_state.assert_called_once_with(state)
+    assert result is mock_result_state
+
+
+def test_entity_tracking_flow_with_mixed_dependencies():
+    """Test flow with mixed explicit and container dependencies."""
+    # Arrange - Create mocks and container
+    mock_entity_service = Mock(spec=EntityService)
+    mock_container = Mock()
+    
+    # Configure container to return mocks for other dependencies
+    mock_container.get.side_effect = lambda name: {
+        "entity_tracker_tool": Mock(spec=EntityTracker),
+        "entity_extractor_tool": Mock(spec=EntityExtractor),
+        "context_analyzer_tool": Mock(spec=ContextAnalyzer),
+        "entity_resolver_tool": Mock(spec=EntityResolver),
+    }.get(name)
+    
+    # Setup mock response
+    mock_result_state = Mock(spec=EntityTrackingState)
+    mock_entity_service.process_article_with_state.return_value = mock_result_state
+    
+    # Create test state
+    state = EntityTrackingState(
+        article_id=1,
+        content="Test content",
+        title="Test Article",
+        published_at=datetime(2025, 1, 1)
+    )
+    
+    # Act - Initialize flow with explicit entity_service and container for other deps
+    flow = EntityTrackingFlow(
+        container=mock_container,
+        entity_service=mock_entity_service
+    )
     
     # Call process method
-    result = flow.process(mock_state)
+    result = flow.process(state)
     
-    # Verify service method was called
-    mock_entity_service.process_article_with_state.assert_called_once_with(mock_state)
+    # Assert - Verify explicit dependency was used
+    mock_entity_service.process_article_with_state.assert_called_once_with(state)
     assert result is mock_result_state
+    
+    # Container should not be called for explicitly provided dependencies
+    for call in mock_container.get.call_args_list:
+        assert call[0][0] != "entity_service"
 
 
-def test_process_new_articles_method():
-    """Test the process_new_articles method."""
-    # Setup mocks
+def test_entity_tracking_flow_handles_errors():
+    """Test error handling in the flow."""
+    # Arrange
     mock_entity_service = Mock(spec=EntityService)
-    mock_result_state = Mock(spec=EntityBatchTrackingState)
-    mock_entity_service.process_articles_batch.return_value = mock_result_state
+    mock_entity_service.process_article_with_state.side_effect = Exception("Test error")
     
-    # Initialize flow
+    # Create test state
+    state = EntityTrackingState(
+        article_id=1,
+        content="John Doe visited the city.",
+        title="Test Article",
+        published_at=datetime(2025, 1, 1)
+    )
+    
+    # Create flow with mock service
     flow = EntityTrackingFlow(entity_service=mock_entity_service)
     
-    # Call process_new_articles method
-    result = flow.process_new_articles()
+    # Act - The flow should catch the exception and return the state with error
+    result_state = flow.process(state)
     
-    # Verify service method was called with correct state
-    mock_entity_service.process_articles_batch.assert_called_once()
-    # Verify the state passed to process_articles_batch is EntityBatchTrackingState with status_filter="analyzed"
-    called_state = mock_entity_service.process_articles_batch.call_args[0][0]
-    assert isinstance(called_state, EntityBatchTrackingState)
-    assert called_state.status_filter == "analyzed"
-    assert result is mock_result_state
+    # Assert
+    mock_entity_service.process_article_with_state.assert_called_once_with(state)
+    assert result_state.status == TrackingStatus.FAILED
+    assert "Test error" in result_state.error_details.message
+    assert result_state.error_details.task == "entity_processing"
 
 
-@patch("local_newsifier.flows.entity_tracking_flow.article_crud")
-def test_process_article_method(mock_article_crud):
-    """Test the process_article method (legacy)."""
-    # Setup mocks
-    mock_entity_service = Mock()  # Don't use spec to avoid attribute constraints
-    mock_session = Mock()
-    mock_article = Mock()
-    mock_article.id = 123
-    mock_article.content = "Test content"
-    mock_article.title = "Test title"
-    mock_article.published_at = datetime.now(timezone.utc)
+def test_dependency_caching():
+    """Test that dependencies are cached after being loaded."""
+    # Arrange - Create mock container with tracking
+    mock_container = MagicMock()
     
-    # Setup session context manager mock properly
-    mock_context_manager = MagicMock()
-    mock_context_manager.__enter__.return_value = mock_session
-    mock_context_manager.__exit__.return_value = None
-    mock_entity_service.session_factory.return_value = mock_context_manager
+    # Track calls to specific dependencies
+    call_counts = {}
     
-    # Configure article crud mock
-    mock_article_crud.get.return_value = mock_article
+    def track_gets(key):
+        call_counts[key] = call_counts.get(key, 0) + 1
+        if key == "entity_service":
+            return Mock(spec=EntityService)
+        return Mock()
     
-    # Configure result state
-    mock_result_state = Mock(spec=EntityTrackingState)
-    mock_result_state.entities = [{"entity": "test"}]
-    mock_entity_service.process_article_with_state.return_value = mock_result_state
+    mock_container.get.side_effect = track_gets
     
-    # Initialize flow
-    flow = EntityTrackingFlow(entity_service=mock_entity_service)
+    # Act - Create flow
+    flow = EntityTrackingFlow(container=mock_container)
     
-    # Call process_article method
-    result = flow.process_article(article_id=123)
+    # Access dependencies multiple times
+    service1 = flow.entity_service
+    service2 = flow.entity_service
+    tracker1 = flow.entity_tracker
+    tracker2 = flow.entity_tracker
     
-    # Verify article was retrieved
-    mock_article_crud.get.assert_called_once_with(mock_session, id=123)
+    # Assert - Each dependency should only be loaded once
+    assert call_counts.get("entity_service", 0) == 1
+    assert call_counts.get("entity_tracker_tool", 0) == 1
     
-    # Verify process was called with correct state
-    mock_entity_service.process_article_with_state.assert_called_once()
-    called_state = mock_entity_service.process_article_with_state.call_args[0][0]
-    assert isinstance(called_state, EntityTrackingState)
-    assert called_state.article_id == 123
-    
-    # Verify result
-    assert result == [{"entity": "test"}]
-
-
-def test_get_entity_dashboard_method():
-    """Test the get_entity_dashboard method."""
-    # Setup mocks
-    mock_entity_service = Mock(spec=EntityService)
-    mock_result_state = Mock(spec=EntityDashboardState)
-    mock_result_state.dashboard_data = {"dashboard": "data"}
-    mock_entity_service.generate_entity_dashboard.return_value = mock_result_state
-    
-    # Initialize flow
-    flow = EntityTrackingFlow(entity_service=mock_entity_service)
-    
-    # Call get_entity_dashboard method
-    result = flow.get_entity_dashboard(days=30, entity_type="PERSON")
-    
-    # Verify service method was called with correct state
-    mock_entity_service.generate_entity_dashboard.assert_called_once()
-    called_state = mock_entity_service.generate_entity_dashboard.call_args[0][0]
-    assert isinstance(called_state, EntityDashboardState)
-    assert called_state.days == 30
-    assert called_state.entity_type == "PERSON"
-    
-    # Verify result
-    assert result == {"dashboard": "data"}
-
-
-def test_find_entity_relationships_method():
-    """Test the find_entity_relationships method."""
-    # Setup mocks
-    mock_entity_service = Mock(spec=EntityService)
-    mock_result_state = Mock(spec=EntityRelationshipState)
-    mock_result_state.relationship_data = {"relationship": "data"}
-    mock_entity_service.find_entity_relationships.return_value = mock_result_state
-    
-    # Initialize flow
-    flow = EntityTrackingFlow(entity_service=mock_entity_service)
-    
-    # Call find_entity_relationships method
-    result = flow.find_entity_relationships(entity_id=456, days=15)
-    
-    # Verify service method was called with correct state
-    mock_entity_service.find_entity_relationships.assert_called_once()
-    called_state = mock_entity_service.find_entity_relationships.call_args[0][0]
-    assert isinstance(called_state, EntityRelationshipState)
-    assert called_state.entity_id == 456
-    assert called_state.days == 15
-    
-    # Verify result
-    assert result == {"relationship": "data"}
+    # Dependencies should be identical (cached)
+    assert service1 is service2
+    assert tracker1 is tracker2
